feat: save hotel infos, rooms and vacancies left in cache with redis
Changes to save hotel infos, total vacancies left and rooms infos in cache. Now, if we have this infos en cache, they will be sent, if not, will get this infos from db and save in cache. If an user create or change your reservation, the cache will be deleted to update with new infos